+++ plugin/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## `0.0.4`
+
+* `eth-state-trie` and `eth-account-snapshot` support.
+  * Accepts Raw RLP encoded input.
+
 ## `0.0.3`
 
 * Improve in-code docs for the plugin loader.